external query file
