+++ package.json
@@ -16,6 +16,7 @@
     "@adyen/adyen-web": "^3.18.2",
     "@babel/plugin-proposal-optional-chaining": "^7.12.1",
     "@commercetools/sdk-auth": "^3.0.10",
+    "@vue/apollo-composable": "^4.0.0-alpha.12",
     "@vue/composition-api": "^1.0.0-rc.5",
     "apollo-cache-inmemory": "^1.6.6",
     "apollo-client": "^2.6.10",
